# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "MeshCore network observation backend. Ingests LoRa packets from MQTT brokers, stores in PostgreSQL, and streams live events via WebSocket.",
        "title": "MeshCore Beacon API",
        "termsOfService": "https://github.com/MeshCore-Beacon/beacon-server",
        "contact": {
            "name": "MeshCore Beacon",
            "url": "https://github.com/MeshCore-Beacon/beacon-server"
        },
        "license": {
            "name": "AGPL-3-or-later"
        },
        "version": "1.4.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/channels": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "summary": "List channels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Single-byte channel hash (hex)",
                        "name": "hash",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by IATA code (case-insensitive)",
                        "name": "iata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by IATA code(s), comma-separated e.g. YVR,YYJ",
                        "name": "iatas",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by region ID, expands to member IATAs",
                        "name": "regionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by region slug, expands to member IATAs",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "last_seen epoch ms of last item for pagination",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ChannelSummaryPage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/channels/{channelID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "summary": "Get channel detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Channel integer ID",
                        "name": "channelID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Channel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/channels/{channelID}/messages": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "summary": "List messages for a channel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Channel integer ID",
                        "name": "channelID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Return messages after this epoch ms",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by IATA code(s), comma-separated e.g. YVR or YVR,YYJ",
                        "name": "iatas",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by region ID, expands to member IATAs",
                        "name": "regionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by region slug, expands to member IATAs",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by transport scope name e.g. %23bc (URL-encoded #bc)",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Message ID of last item for pagination (results ordered newest first)",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
...
        "/messages": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "List channel messages",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by channel integer ID (mutually exclusive with channelHash)",
                        "name": "channelID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by channel hash byte hex (mutually exclusive with channelID)",
                        "name": "channelHash",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Return messages after this epoch ms",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by IATA code(s), comma-separated e.g. YVR or YVR,YYJ",
                        "name": "iatas",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by region ID, expands to member IATAs",
                        "name": "regionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by region slug, expands to member IATAs",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by transport scope name e.g. %23bc (URL-encoded #bc)",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Message ID of last item for pagination (results ordered newest first)",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.Channel": {
            "type": "object",
            "properties": {
                "channelHash": {
                    "description": "hex-encoded single-byte hash",
                    "type": "string"
                },
                "hashtag": {
                    "description": "tag name without # prefix; non-nil only for hashtag channels",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isHashtag": {
                    "description": "true if key was derived from a hashtag PSK",
                    "type": "boolean"
                },
                "keyFingerprint": {
                    "description": "first 8 bytes of SHA256(key), hex-encoded",
                    "type": "string"
                },
                "keyKnown": {
                    "description": "true if Beacon has a decryption key for this channel",
                    "type": "boolean"
                },
                "lastSeen": {
                    "description": "epoch ms, time of most recent message",
                    "type": "integer"
                },
                "messageCount": {
                    "type": "integer"
                },
                "name": {
                    "description": "display name from config or nil",
                    "type": "string"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.ChannelSummary": {
            "type": "object",
            "properties": {
                "channelHash": {
                    "description": "hex-encoded single-byte hash",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isHashtag": {
                    "description": "true if key was derived from a hashtag PSK",
                    "type": "boolean"
                },
                "keyKnown": {
                    "description": "true if Beacon has a decryption key for this channel",
                    "type": "boolean"
                },
                "lastSeen": {
                    "description": "epoch ms, time of most recent message",
                    "type": "integer"
                },
                "name": {
                    "description": "display name from config or nil",
                    "type": "string"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.ChannelSummaryPage": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ChannelSummary"
                    }
                },
                "nextCursor": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api_handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g. \"not_found\", \"bad_request\"",
                    "type": "string"
                },
                "message": {
                    "description": "e.g. \"channel not found\"",
                    "type": "string"
                }
            }
        },
...
    },
...
}